added dynamic routing for the categories
