--- server/api/tours/index.get.ts
@@ -1,11 +1,0 @@
-import Categories from "~/server/models/categories"
-
-
-export default defineEventHandler(async (event) => {
-
-    const categories = await Categories.find();
-
-    return {
-        categories
-    }
-})
